+++ src/renderer/pages/settings/page.tsx
@@ -1,0 +1,8 @@
+// Components
+import { useEffect, useState } from 'react';
+import { Container, FlexDiv } from 'renderer/components/container';
+import { CFTextIcon } from 'renderer/components/icons/codeforces';
+
+// Tools
+import { classNames } from 'renderer/tools/css_tools';
+import { Link } from 'react-router-dom';